Rename ITutorial to ITutorialRepository
Better understanding of the file
