+++ app/models/repositories/tutorial/repoTutorials.py
@@ -1,5 +1,5 @@
 from .tutorial import Tutorial
-from .ITutorial import ITutorialRepository
+from .ITutorialRepository import ITutorialRepository
 
 class Tutorial_mock_repo(ITutorialRepository):
     def __init__(self):
